Revert "configure-alertmanager-operator test & refactor"
