--- test/operators/configurealertmanager.go
@@ -1,31 +1,0 @@
-package operators
-
-import (
-	"github.com/onsi/ginkgo"
-	"github.com/openshift/osde2e/pkg/helper"
-)
-
-var _ = ginkgo.Describe("[OSD] Configure AlertManager Operator", func() {
-	var operatorName = "configure-alertmanager-operator"
-	var operatorNamespace string = "openshift-monitoring"
-	var operatorLockFile string = "configure-alertmanager-operator-lock"
-	var defaultDesiredReplicas int32 = 1
-
-	var clusterRoles = []string{
-		"configure-alertmanager-operator",
-	}
-
-	var clusterRoleBindings = []string{}
-
-	var roleBindings = []string{
-		"configure-alertmanager-operator",
-	}
-
-	h := helper.New()
-	checkClusterServiceVersion(h, operatorNamespace, operatorName)
-	checkConfigMapLockfile(h, operatorNamespace, operatorLockFile)
-	checkDeployment(h, operatorNamespace, operatorName, defaultDesiredReplicas)
-	checkClusterRoles(h, clusterRoles)
-	checkClusterRoleBindings(h, clusterRoleBindings)
-	checkRoleBindings(h, operatorNamespace, roleBindings)
-})
